refactor(rskim-compress): consolidate log tests, remove dead code (#304/#327)
Completes the half-move from prior phase:

- Remove `#![allow(dead_code)]` blanket from crates/rskim/src/cmd/log.rs
- Delete dead log-parsing/compression helpers (was ~1700 lines of dead code
  that duplicated the live implementation in rskim-compress::log)
- Keep CLI glue only: parse_flags, run, record_analytics, compress_log
  delegation wrapper, convert_log_result, emit_result (441 lines total)
- Port 54 logic tests to crates/rskim-compress/src/log.rs (9 sub-modules:
  tier_detection, type, debug_filter, dedup, timestamp, json_field,
  stack_trace, python_continuation, python_traceback)
- Add self-contained fixture dir: crates/rskim-compress/tests/fixtures/log/
  (9 fixture files copied from crates/rskim/tests/fixtures/cmd/log/)
- 8 CLI glue tests remain in cmd/log.rs (flag_tests + delegation_tests)
- Apply #[allow(clippy::unwrap_used, ...)] between #[cfg(test)] and mod tests
  to satisfy crate-level lint deny in rskim-compress (matches rskim-contract
  pattern)

Verification:
- cargo nextest run -p rskim-compress -j 4: 59/59 passed
- cargo test -p rskim-compress --doc: ok
- cargo clippy -p rskim-compress --all-targets -- -D warnings: clean
- cargo clippy -p rskim --all-targets -- -D warnings: clean
- cargo fmt -- --check: clean

Note: cargo nextest run -p rskim -E 'test(/log/)' cannot be run in this
environment — the rskim::bin/skim test binary hangs during dyld startup
(heavy tree-sitter/SQLite/ring dep loading under load; pre-existing
machine-specific issue documented in MEMORY.md, not caused by this change).
