+++ crates/rskim-compress/tests/fixtures/log/dedup_error_warn.txt
@@ -1,0 +1,5 @@
+ERROR: connection refused
+WARN: connection refused
+ERROR: connection refused
+WARN: connection refused
+INFO: retrying
