+++ crates/rskim-compress/tests/fixtures/log/stack_trace_python.txt
@@ -1,0 +1,10 @@
+ERROR: ValueError: invalid literal for int()
+  File "/app/service.py", line 55, in parse_value
+    result = int(value)
+  File "/app/service.py", line 102, in handle_request
+    return parse_value(data)
+  File "/app/service.py", line 201, in run
+    handle_request(payload)
+  File "/usr/lib/python3.11/threading.py", line 1000, in _bootstrap
+    self.run()
+INFO: continuing after error
